Updates
Created the 3 other pages for the multi page layout, created a "more info" button in the about section that links to the about.html page, embedded the "see profile" buttons inside a card footer in each card for the contact section, changed the border color around the profile images to the same color as the border around the cards. Removed the second About Us paragraph on the single page layout and added a "read more" button to expand the section to display the second paragraph.